Set the return type of evaluate to int

+++ src/edu/rlv/cosc60/activities/PostFixExpression.java
@@ -14,7 +14,7 @@ public class PostFixExpression {
         throw new UnsupportedOperationException();
     }
     
-    public static float evaluate(String postFix){
+    public static int evaluate(String postFix){
         throw new UnsupportedOperationException();
     }
